+++ tests/testthat/test_io_visualization.R
@@ -219,7 +219,7 @@ test_that("Seurat plotting helpers return ggplot objects", {
   expect_equal(length(multi_feature_plot$guides$guides), 1)
   expected_geom <- if (rlang::is_installed("shadowtext")) "GeomShadowText" else "GeomText"
   expect_identical(class(dim_plot_repel$layers[[2]]$geom)[[1]], expected_geom)
-  expect_equal(as.numeric(dim_plot$theme$legend.text$margin[[4]]), -4)
+  expect_equal(as.numeric(dim_plot$theme$legend.text$margin[[4]]), 2)
   expect_true(all(vapply(
     c(list(multi_feature_plot), multi_feature_plot$patches$plots),
     function(plot) is.language(plot$labels$title),